#8 df: Refactoring. Created own Parser file
