+++ src/components/scenariotest/ScenarioTestMenu.jsx
@@ -9,6 +9,7 @@ import axios from 'axios';
 import pluralize from 'pluralize';
 import compromise from 'compromise';
 import ActivityValidator from './ActivityValidator';
+import ActivityParser from "./ActivityParser.jsx";
 
 export default function ScenarioTestMenu(props) {
 
@@ -127,17 +128,14 @@ export default function ScenarioTestMenu(props) {
         colorActiveActivities(event);
 
         // update the view for the selected edge
-        let newSelectedActivity = allActivities.find((artifact) => artifact.description === event.target.value);
+        let selectedActivity = allActivities.find((artifact) => artifact.description === event.target.value);
 
         // validate activity
-        let activityDescription = newSelectedActivity.description;
-        let allActiveEdges = props.edges.filter((edge) => edge.name === activityDescription);
-        let allElements = findAllElements(allActiveEdges);
-        let wordArray = buildWordArray(allElements);
+        let wordArray = ActivityParser(props.nodes, props.edges, selectedActivity);
         let isValidActivity = ActivityValidator(wordArray);
 
         let allDefinedScenariosCopy = deepCopy(allDefinedScenarios);
-        allDefinedScenariosCopy[index].activity = newSelectedActivity;
+        allDefinedScenariosCopy[index].activity = selectedActivity;
         allDefinedScenariosCopy[index].selected_mode = null;
         allDefinedScenariosCopy[index].generatedScenariosList = null;
         allDefinedScenariosCopy[index].filteredScenariosList = null;
@@ -359,10 +357,7 @@ export default function ScenarioTestMenu(props) {
     };
 
     const getScenariosForActivityAndMode = (activity, mode) => {
-        let activityDescription = activity.description;
-        let allActiveEdges = props.edges.filter((edge) => edge.name === activityDescription);
-        let allElements = findAllElements(allActiveEdges);
-        let wordArray = buildWordArray(allElements);
+        let wordArray = ActivityParser(props.nodes, props.edges, activity);
 
         let generatedSentences;
         if(mode === "What if") {
@@ -374,103 +369,6 @@ export default function ScenarioTestMenu(props) {
 
         //console.log(generatedSentences);
         return generatedSentences;
-    }
-
-    //TODO: Add annotations
-    const findAllElements = (edges) => {
-        const allElements = [];
-        let currentSource = null;
-
-        //finding the first element
-        for (const edge of edges) {
-            const source = edge.source;
-            let isFirstElement = true;
-
-            for (const otherEdge of edges) {
-                const otherSource = otherEdge.source;
-                const otherTarget = otherEdge.target;
-
-                if (source !== otherSource && source === otherTarget) {
-                    isFirstElement = false;
-                    break;
-                }
-            }
-
-            if (isFirstElement) {
-                currentSource = source;
-                break;
-            }
-        }
-
-        while (currentSource !== null) {
-            let found = false;
-
-            for (const edge of edges) {
-                const source = edge.source;
-                const target = edge.target;
-
-                if (currentSource === source) {
-                    found = true;
-                    const node = props.nodes.find((n) => n.id === currentSource);
-                    allElements.push(node);
-                    allElements.push(edge);
-                    currentSource = target;
-                    break;
-                }
-            }
-
-            if (!found) {
-                break;
-            }
-        }
-
-        //finding the last element
-        for (const edge of edges) {
-            const target = edge.target;
-
-            if (currentSource === target) {
-                const node = props.nodes.find((n) => n.id === target);
-                allElements.push(node);
-            }
-        }
-
-        return allElements.reverse();
-    }
-
-    const buildWordArray = (allElements) => {
-        let sentenceArray = [];
-        for (const element of allElements) {
-            // if element is not an edge
-            if (element.data !== undefined) {
-                let typeString = null;
-                if (element.data.icon === "Document") {
-                    typeString = "work object";
-                }
-                else if (element.data.icon === "Person") {
-                    typeString = "person";
-                }
-                //TODO: Edit Annotation. Maybe it is not necessary
-                else if (element.data.icon === "Annotation") {
-                    typeString = "annotation";
-                }
-                else if (element.data.icon === "System") {
-                    typeString = "system";
-                }
-                let wordObject = {name: element.data.label.toLowerCase(), type: typeString};
-                sentenceArray.push(wordObject);
-            }
-            // else element is an edge
-            else {
-                let name = element.label.endsWith("s") ? element.label.slice(0, -1) : element.label;
-                let typeString = "verb";
-                if (name === "in") {
-                    typeString = "preposition"
-                }
-                let wordObject = {name: name.toLowerCase(), type: typeString};
-                sentenceArray.push(wordObject);
-            }
-        }
-        return sentenceArray;
     }
 
     const generateScenarios = (wordArray, mode) => {
